requirements.txt
numpy
pandas
logging
setuptools
google-cloud-storage
scikit-learn
pyyaml